Standard Deviation of the Relative Range

Compute the constant \(d_3\), the standard deviation of the relative range \(W = R / \sigma\) for a sample of size \(n\) drawn from a normal distribution, where \(R\) is the sample range and \(\sigma\) is the process standard deviation.

Usage

d3(n)

Arguments

n

Integer sample size, at least 2. Can be a vector.

Value

Numeric vector of \(d_3\) values, the same length as n.

Details

Under normality, \(W\) has the studentized range distribution with infinite denominator degrees of freedom. Using the identity \(E[W^2] = 2 \int_0^\infty w \{1 - F_W(w)\} dw\), the standard deviation is obtained as \(d_3 = \sqrt{E[W^2] - d_2^2}\), where \(d_2\) is computed by d2. The integrals are evaluated numerically using ptukey with infinite degrees of freedom.

References

Montgomery, D. C. (2009). Introduction to Statistical Quality Control, 6th ed. Wiley.

Barbosa, E. P., Gneri, M. A. and Meneguetti, A. (2013). Range control charts revisited: Simpler Tippett-like formulae, its practical implementation, and the study of false alarm. Communications in Statistics - Simulation and Computation, 42(2), 247–262. doi:10.1080/03610918.2011.639967 .

See also

Author

Daniela R. Recchia, Emanuel P. Barbosa

Examples

d3(7)
#> [1] 0.8332053
d3(2:10)
#> [1] 0.8525025 0.8883680 0.8798082 0.8640819 0.8480397 0.8332053 0.8198311
#> [8] 0.8078343 0.7970507