Plot a DS-np Control Chart

S3 method for plotting objects of class "cchart.DSnp".

# S3 method for class 'cchart.DSnp'
plot(x, ...)

Arguments

x

An object of class "cchart.DSnp".

...

Additional graphical parameters (currently unused).

Value

Invisible x.

Details

The plot shows the first-stage counts x1 against the sample index. Points requiring a second sample are shown as open circles. Points that signal (at either stage) are shown in red. Horizontal lines mark the fractional limits.

Note that ucl2 is on the scale of the combined count x1 + x2. The plot is a simple operational visualization; the full decision logic is in the returned object.