Table of values for the constants d2, d3 and c4.

This function is used to build a table of values for the constants d2, d3 and c4 for sucessive values of sample size n.

table.const(n)

Arguments

n
The maximum size.

Value

Return the values of these three constants.

Details

It builds a table in matrix form with 3 columns (one for each constant) and one row for each value of n from 2 to a specified value.

See also

d2,d3,c4

Examples

table.const(17)

#>          d2        d3        c4
#> 2  1.128379 0.8525025 0.7978846
#> 3  1.692569 1.5429670 0.8862269
#> 4  2.058751 1.9337212 0.9213177
#> 5  2.325929 2.2098288 0.9399856
#> 6  2.534413 2.4226390 0.9515329
#> 7  2.704357 2.5950987 0.9593688
#> 8  2.847201 2.7396049 0.9650305
#> 9  2.970026 2.8636363 0.9693107
#> 10 3.077505 2.9720515 0.9726593
#> 11 3.172873 3.0681828 0.9753501
#> 12 3.258455 3.1544127 0.9775594
#> 13 3.335980 3.2325016 0.9794056
#> 14 3.406763 3.3037856 0.9809714
#> 15 3.471827 3.3693023 0.9823162
#> 16 3.531983 3.4298725 0.9834835
#> 17 3.587884 3.4861562 0.9845064